% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_data.R
\name{fit_data}
\alias{fit_data}
\title{Fitting right censored survival data to distribution}
\usage{
fit_data(data, dist, time = "Time", censor = "Censor", by = "")
}
\arguments{
\item{data}{a dataframe containing a time column and a censor column.}

\item{dist}{a string name for a distribution that has a corresponding desnity function and a distribution function.}

\item{time}{the string name of the time column of the dataframe. defaults to "time".}

\item{censor}{the string name of the censor column of the dataframe. defaults to "censor". the censor column must be 
a numeric indicator variable where complete times correspond to a value of 1 and incomplete times correspond to 0.}

\item{by}{the string name of a grouping variable. if specified, the function returns a list.}
}
\description{
Fits right censored data to a distribution using maximum liklihood estimates.
}
\seealso{
\code{\link[fitdistrplus]{fitdistcens()}}
}
