% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/placeholdR.R
\docType{package}
\name{placeholdR}
\alias{placeholdR}
\alias{placeholdR-package}
\title{foo: A package for computating the notorious bar statistic.}
\description{
Executes parametric survival analysis techniques similar to those in Minitab. Fits right censored data to a
given parametric distribution, produces summary statistics of the fitted distribution, and plots parametric survival, 
hazard, and cumulative hazard plots. Produces Anderson-Darling test statistic and probability plots to assess goodness
of fit of right censored data to a distribution.
}
