% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_haz.R
\name{plot_haz}
\alias{plot_haz}
\title{Plotting parametric hazard curves}
\usage{
plot_haz(data, dist, time = "Time", censor = "Censor", by = "")
}
\arguments{
\item{data}{a dataframe containing a time column and a censor column.}

\item{dist}{a string name for a distribution that has a corresponding desnity function and distribution function.}

\item{time}{the string name of the time column of the dataframe. defaults to "time".}

\item{censor}{the string name of the censor column of the dataframe. defaults to "censor". the censor column must be 
a numeric indicator variable where complete times correspond to a value of 1 and incomplete times correspond to 0.}

\item{by}{the string name of a grouping variable. If specified, multiple lines will be plotted.}
}
\description{
Plots hazard curve of right censored data given that it follows a specified parametric distribution.
}
