% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{prob}
\alias{prob}
\title{Survival probability based on parametric distribution}
\usage{
prob(data, dist, num, lower.tail = F, time = "Time", censor = "Censor")
}
\arguments{
\item{data}{A dataframe containing a time column and a censor column.}

\item{dist}{A string name for a distribution that has a corresponding desnity function and a distribution function.}

\item{num}{A scalar quantity}

\item{lower.tail}{logical; if \code{F} (default), probability is P(T > \code{num}), otherwise, P(T < \code{num}).}

\item{time}{The string name of the time column of the dataframe. Defaults to "Time".}

\item{censor}{The string name of the censor column of the dataframe. Defaults to "Censor". The censor column must be 
a numeric indicator variable where complete times correspond to a value of 1 and incomplete times correspond to 0.}
}
\description{
Compute survival probabilities given that the data follows a specified parametric distribution.
}
\examples{
library(survival) 
data("rats")
prob(rats, "lnorm", 110, time = "time", censor = "status")
}
