% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_summary.R
\name{surv_summary}
\alias{surv_summary}
\title{Summary statistics based on parametric distribution}
\usage{
surv_summary(data, dist, time = "Time", censor = "Censor", by = "")
}
\arguments{
\item{data}{A dataframe containing a time column and a censor column.}

\item{dist}{A string name for a distribution that has a corresponding desnity function and a distribution function.}

\item{time}{The string name of the time column of the dataframe. Defaults to "Time".}

\item{censor}{The string name of the censor column of the dataframe. Defaults to "Censor". The censor column must be 
a numeric indicator variable where complete times correspond to a value of 1 and incomplete times correspond to 0.}

\item{by}{The string name of a grouping variable. If specified, returns summary statistics for each group.}
}
\description{
Estimates various statistics, including median, mean, standard deviation, and 
percentiles of survival time given that the data follows a specified parametric distribution.
}
